Added MenuManager.cs , IMenuManager.cs , SQLiteConnection.cs, Edited UserList.cs
1. UserSchedule.Services Namespace - MenuManager Class:
The MenuManager class implements the IMenuManager interface.
The class initializes an SQLite database connection on its creation and ensures the tables for User, UserList, and UserListItem are created.
Provides methods to:
Add new entities like User, UserList, and UserListItem.
Commit changes to the SQLite database.
Delete lists and items based on their IDs.
Retrieve a UserList or UserListItem by their IDs.

2. menu.Services Namespace - IMenuManager Interface:
Defines the contract for a menu manager service.
Specifies methods to add a User, UserList, and UserListItem.

3. menu Namespace - SQLiteDb Static Class:
Provides a utility to get an SQLite connection.
Determines the path for the SQLite database file named ListManager.db within the application's data directory.
In summary, today's work mainly revolved around the establishment of services to manage users, lists, and list items with functionalities for addition, deletion, and retrieval. An SQLite database has been employed for data storage, with utility functions ensuring smooth database operations. The defined interface ensures that the service adheres to specific operations for managing the menu.