Local database (#3)
* Added MenuManager.cs , IMenuManager.cs , SQLiteConnection.cs, Edited UserList.cs

1. UserSchedule.Services Namespace - MenuManager Class:
The MenuManager class implements the IMenuManager interface.
The class initializes an SQLite database connection on its creation and ensures the tables for User, UserList, and UserListItem are created.
Provides methods to:
Add new entities like User, UserList, and UserListItem.
Commit changes to the SQLite database.
Delete lists and items based on their IDs.
Retrieve a UserList or UserListItem by their IDs.

2. menu.Services Namespace - IMenuManager Interface:
Defines the contract for a menu manager service.
Specifies methods to add a User, UserList, and UserListItem.

3. menu Namespace - SQLiteDb Static Class:
Provides a utility to get an SQLite connection.
Determines the path for the SQLite database file named ListManager.db within the application's data directory.
In summary, today's work mainly revolved around the establishment of services to manage users, lists, and list items with functionalities for addition, deletion, and retrieval. An SQLite database has been employed for data storage, with utility functions ensuring smooth database operations. The defined interface ensures that the service adheres to specific operations for managing the menu.

* Add AppConfiguration.cs, AppConfiguration.cs, Add foreign keys and initialize database links

1. AppConfiguration. cs defines the AppConfiguration class as an accessor for application configuration, specifically for database configuration. By implementing the IAppConfiguration interface, the class provides methods and properties to obtain the connection string and complete database name (including path) of the database. It reads configuration information, such as database names and connection strings, from external configuration sources such as appsettings. json, and provides default values as an alternative to missing configurations. In this way, other code can access database configuration information in a unified and simplified way through instances of this class.
2. I found that the UserList and UserListItem classes do not contain foreign keys connected to other tables, so I added them

* Cherry pick changes specific to local db branch

---------

Co-authored-by: ccbrowndev <46230726+ccbrowndev@users.noreply.github.com>